refactor : component -> ui 로 수정했습니다
